admin_app/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework==3.14.0
Faker==18.6.2
idna==3.4
psycopg==3.1.9
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
typing_extensions==4.5.0
urllib3==2.0.2
